bug fix - add coin to correct person and refactorings
Refactored the `WasCorrectlyAnswered` method in `Game.cs` to simplify the logic for handling players in the penalty box by removing nested `if` statements. Renamed `DidPlayerWin` method in `Player.cs` to `YetToWin` to better reflect its purpose. Corrected a typo in the console output message from "Answer was corrent!!!!" to "Answer was correct!!!!". Updated test approval files to reflect these changes.